restructure article-structure plugin a bit
